+++ core/src/dird/dir_plugins.cc
@@ -209,7 +209,8 @@ bRC GeneratePluginEvent(JobControlRecord* jcr,
   plugin_ctx_list = jcr->plugin_ctx_list;
   event.eventType = eventType;
 
-  Dmsg2(debuglevel, "dir-plugin_ctx_list=%p JobId=%d\n", plugin_ctx_list,
+  Dmsg2(debuglevel, "dir-plugin_ctx_list=%p JobId=%" PRIu32 "\n",
+        plugin_ctx_list,
         jcr->JobId);
 
   // See if we need to trigger the loaded plugins in reverse order.
@@ -248,7 +249,7 @@ void DumpDirPlugin(Plugin* plugin, FILE* fp)
   if (!plugin) { return; }
 
   info = (PluginInformation*)plugin->plugin_information;
-  fprintf(fp, "\tversion=%d\n", info->version);
+  fprintf(fp, "\tversion=%" PRIu32 "\n", info->version);
   fprintf(fp, "\tdate=%s\n", NPRTB(info->plugin_date));
   fprintf(fp, "\tmagic=%s\n", NPRTB(info->plugin_magic));
   fprintf(fp, "\tauthor=%s\n", NPRTB(info->plugin_author));
@@ -327,10 +328,15 @@ static bool IsPluginCompatible(Plugin* plugin)
   }
   if (info->version != DIR_PLUGIN_INTERFACE_VERSION) {
     Jmsg(NULL, M_ERROR, 0,
-         T_("Plugin version incorrect. Plugin=%s wanted=%d got=%d\n"),
-         plugin->file, DIR_PLUGIN_INTERFACE_VERSION, info->version);
-    Dmsg3(50, "Plugin version incorrect. Plugin=%s wanted=%d got=%d\n",
-          plugin->file, DIR_PLUGIN_INTERFACE_VERSION, info->version);
+         T_("Plugin version incorrect. Plugin=%s wanted=%" PRIu32
+            " got=%" PRIu32 "\n"),
+         plugin->file, static_cast<uint32_t>(DIR_PLUGIN_INTERFACE_VERSION),
+         info->version);
+    Dmsg3(50,
+          "Plugin version incorrect. Plugin=%s wanted=%" PRIu32
+          " got=%" PRIu32 "\n",
+          plugin->file, static_cast<uint32_t>(DIR_PLUGIN_INTERFACE_VERSION),
+          info->version);
     return false;
   }
   if (!Bstrcasecmp(info->plugin_license, "Bareos AGPLv3")
@@ -366,7 +372,7 @@ static inline PluginContext* instantiate_plugin(JobControlRecord* jcr,
   b_ctx->jcr = jcr;
   b_ctx->plugin = plugin;
 
-  Dmsg2(debuglevel, "Instantiate dir-plugin_ctx_list=%p JobId=%d\n",
+  Dmsg2(debuglevel, "Instantiate dir-plugin_ctx_list=%p JobId=%" PRIu32 "\n",
         jcr->plugin_ctx_list, jcr->JobId);
 
   ctx = (PluginContext*)malloc(sizeof(PluginContext));
@@ -435,7 +441,8 @@ void DispatchNewPluginOptions(JobControlRecord* jcr)
 
       if (instance > HIGHEST_PLUGIN_INSTANCE) {
         Jmsg(NULL, M_ERROR, 0,
-             T_("Illegal DIR plugin options encountered, %s instance %d "
+             T_("Illegal DIR plugin options encountered, %s instance %" PRIu32
+                " "
                 "skipping\n"),
              plugin_options, instance);
         continue;
@@ -504,7 +511,7 @@ void FreePlugins(JobControlRecord* jcr)
 {
   if (!dird_plugin_list || !jcr->plugin_ctx_list) { return; }
 
-  Dmsg2(debuglevel, "Free instance dir-plugin_ctx_list=%p JobId=%d\n",
+  Dmsg2(debuglevel, "Free instance dir-plugin_ctx_list=%p JobId=%" PRIu32 "\n",
         jcr->plugin_ctx_list, jcr->JobId);
   for (auto* ctx : jcr->plugin_ctx_list) {
     // Free the plugin instance
@@ -545,7 +552,8 @@ static bRC bareosGetValue(PluginContext* ctx, brDirVariable var, void* value)
     switch (var) {
       case bDirVarJobId:
         *((int*)value) = jcr->JobId;
-        Dmsg1(debuglevel, "dir-plugin: return bDirVarJobId=%d\n", jcr->JobId);
+        Dmsg1(debuglevel, "dir-plugin: return bDirVarJobId=%" PRIu32 "\n",
+              jcr->JobId);
         break;
       case bDirVarJobName:
         *((char**)value) = jcr->Job;
@@ -579,7 +587,8 @@ static bRC bareosGetValue(PluginContext* ctx, brDirVariable var, void* value)
                  sizeof(pr.Name));
         if (!jcr->db->GetPoolRecord(jcr, &pr)) { retval = bRC_Error; }
         *((int*)value) = pr.NumVols;
-        Dmsg1(debuglevel, "dir-plugin: return bDirVarNumVols=%d\n", pr.NumVols);
+        Dmsg1(debuglevel, "dir-plugin: return bDirVarNumVols=%" PRIu32 "\n",
+              pr.NumVols);
         break;
       }
       case bDirVarPool:
@@ -656,22 +665,22 @@ static bRC bareosGetValue(PluginContext* ctx, brDirVariable var, void* value)
         break;
       case bDirVarJobErrors:
         *((int*)value) = jcr->JobErrors;
-        Dmsg1(debuglevel, "dir-plugin: return bDirVarErrors=%d\n",
+        Dmsg1(debuglevel, "dir-plugin: return bDirVarErrors=%" PRIu32 "\n",
               jcr->JobErrors);
         break;
       case bDirVarJobFiles:
         *((int*)value) = jcr->JobFiles;
-        Dmsg1(debuglevel, "dir-plugin: return bDirVarFiles=%d\n",
+        Dmsg1(debuglevel, "dir-plugin: return bDirVarFiles=%" PRIu32 "\n",
               jcr->JobFiles);
         break;
       case bDirVarSDJobFiles:
         *((int*)value) = jcr->dir_impl->SDJobFiles;
-        Dmsg1(debuglevel, "dir-plugin: return bDirVarSDFiles=%d\n",
+        Dmsg1(debuglevel, "dir-plugin: return bDirVarSDFiles=%" PRIu32 "\n",
               jcr->dir_impl->SDJobFiles);
         break;
       case bDirVarSDErrors:
         *((int*)value) = jcr->dir_impl->SDErrors;
-        Dmsg1(debuglevel, "dir-plugin: return bDirVarSDErrors=%d\n",
+        Dmsg1(debuglevel, "dir-plugin: return bDirVarSDErrors=%" PRIu32 "\n",
               jcr->dir_impl->SDErrors);
         break;
       case bDirVarFDJobStatus:
@@ -686,7 +695,7 @@ static bRC bareosGetValue(PluginContext* ctx, brDirVariable var, void* value)
         break;
       case bDirVarLastRate:
         *((int*)value) = jcr->LastRate;
-        Dmsg1(debuglevel, "dir-plugin: return bDirVarLastRate=%d\n",
+        Dmsg1(debuglevel, "dir-plugin: return bDirVarLastRate=%" PRIu32 "\n",
               jcr->LastRate);
         break;
       case bDirVarJobBytes:
