+++ core/src/stored/mac.cc
@@ -153,7 +153,8 @@ static bool CloneRecordInternally(DeviceControlRecord* dcr,
         }
         jcr->setJobType(sos_label.JobType);
         jcr->setJobLevel(sos_label.JobLevel);
-        Dmsg1(200, "joblevel from SOS_LABEL is now %c\n", sos_label.JobLevel);
+        Dmsg1(200, "joblevel from SOS_LABEL is now %c\n",
+              static_cast<int>(sos_label.JobLevel));
 
         // make sure this volume wasn't written by bacula 1.26 or earlier
         ASSERT(sos_label.VerNum >= 11);
@@ -228,7 +229,10 @@ static bool CloneRecordInternally(DeviceControlRecord* dcr,
   }
 
   while (!WriteRecordToBlock(jcr->sd_impl->dcr, jcr->sd_impl->dcr->after_rec)) {
-    Dmsg4(200, "!WriteRecordToBlock blkpos=%u:%u len=%d rem=%d\n", dev->file,
+    Dmsg4(200,
+          "!WriteRecordToBlock blkpos=%u:%u len=%" PRIu32 " rem=%" PRIu32
+          "\n",
+          dev->file,
           dev->block_num, jcr->sd_impl->dcr->after_rec->data_len,
           jcr->sd_impl->dcr->after_rec->remainder);
     if (!jcr->sd_impl->dcr->WriteBlockToDevice()) {
@@ -249,7 +253,8 @@ static bool CloneRecordInternally(DeviceControlRecord* dcr,
   jcr->JobBytes += jcr->sd_impl->dcr->after_rec
                        ->data_len; /* increment bytes of this job */
 
-  Dmsg5(500, "wrote_record JobId=%d FI=%s SessId=%d Strm=%s len=%d\n",
+  Dmsg5(500, "wrote_record JobId=%" PRIu32 " FI=%s SessId=%" PRIu32
+              " Strm=%s len=%" PRIu32 "\n",
         jcr->JobId, FI_to_ascii(buf1, jcr->sd_impl->dcr->after_rec->FileIndex),
         jcr->sd_impl->dcr->after_rec->VolSessionId,
         stream_to_ascii(buf2, jcr->sd_impl->dcr->after_rec->Stream,
@@ -350,7 +355,7 @@ static bool CloneRecordToRemoteSd(DeviceControlRecord* dcr,
 
   // Send a header when needed.
   if (send_header) {
-    if (!sd->fsend("%" PRIu32 " %" PRId32 " 0", rec->FileIndex, rec->Stream)) {
+    if (!sd->fsend("%d %" PRId32 " 0", rec->FileIndex, rec->Stream)) {
       if (!jcr->IsJobCanceled()) {
         Jmsg1(jcr, M_FATAL, 0, T_("Network send error to SD. ERR=%s\n"),
               sd->bstrerror());
@@ -380,7 +385,8 @@ static bool CloneRecordToRemoteSd(DeviceControlRecord* dcr,
   jcr->JobBytes += sd->message_length;
   sd->msg = msgsave;
 
-  Dmsg5(200, "wrote_record JobId=%d FI=%s SessId=%d Strm=%s len=%d\n",
+  Dmsg5(200, "wrote_record JobId=%" PRIu32 " FI=%s SessId=%" PRIu32
+              " Strm=%s len=%" PRIu32 "\n",
         jcr->JobId, FI_to_ascii(buf1, rec->FileIndex), rec->VolSessionId,
         stream_to_ascii(buf2, rec->Stream, rec->FileIndex), rec->data_len);
